Move the Gathering Data Notification as a standalone component.
